Non-threaded version of FST reading.
Based on similar function in fstd2nc.

Code modified to call compute() in parallel over all rows of the current
piece of table.

Note: 'swa' and 'lng' columns had to be re-activated, because they're needed
for the alternative reading function.

'swa' is direct offset in the file where the data is located (in 8-byte units).
'lng' is the length of the data in the file (in 4-byte units).